Isolating ARM specific code in inffast
The NEON specific code will be hosted in the folder
'contrib/optimizations/arm' while the platform independent
C code is hosted in the upper directory.

This allows to easily implement the inffast optimization for other
architectures by simply implementing 2 functions and including the
necessary header in chunk_copy.h (that is used by inflate and inffast).

The idea is with time to move all optimizations to this new folder.

Bug: 769880
Change-Id: I404ec0fdf3f6867c9c124da859ca38bf57b25447
Reviewed-on: https://chromium-review.googlesource.com/740907
Reviewed-by: Chris Blume <cblume@chromium.org>
Commit-Queue: Adenilson Cavalcanti <cavalcantii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#512542}
Signed-off-by: mydongistiny <jaysonedson@gmail.com>