Speed up software CRC-32 computation by a factor of 1.5 to 3
Use the interleaved method of Kadatch and Jenkins in order to make
use of pipelined instructions through multiple ALUs in a single
core. This also speeds up and simplifies the combination of CRCs,
and updates the functions to pre-calculate and use an operator for
CRC combination.

BenzoEdit: Fixed up for ARMv8 CRC32 specific instructions

Change-Id: I6fa4fec4318c94b8e422de3720c5ef8cca69cc77
Signed-off-by: mydongistiny <jaysonedson@gmail.com>